lab3: Split code into modules
